Address PR review comments - refactor layout editor
Refactored layout editor code based on PR feedback:

1. Consolidated ghost preview components
   - Merged GhostMachinePreview and GhostMovePreview into single MachineGhostPreview component
   - Component now accepts machineTypeId, position, rotation, and optional excludeMachineIndex

2. Simplified FloorTileSprite
   - Removed all placement/movement logic
   - Now only dispatches onClick, onHover, onHoverOut events
   - All game logic moved to LayoutPage

3. Extracted selection highlight component
   - Created MachineSelectionHighlight component
   - Removed drawSelectionHighlight from MachineSprite

4. Split LayoutPage into smaller components
   - Created src/components/layout-page/ directory
   - LayoutEditorCanvas.tsx - handles PIXI rendering
   - StorageSection.tsx - storage inventory UI
   - EditorModeIndicator.tsx - mode display and keyboard shortcuts
   - MachineGhostPreview.tsx - ghost preview rendering

5. Fixed React.useEffect import
   - Changed from React.useEffect to direct useEffect import

6. Converted test fixtures from JSON to TypeScript
   - Fixtures now type-checked against GameState
   - Fixed invalid board dimensions (48 -> 8)
   - Fixed operation IDs (crosscut -> cutBoard)
   - Added missing material IDs
   - Removed resolveJsonModule from tsconfig.json

7. Updated test selectors
   - Changed from text= locators to getByText() for better reliability
