+++ src/components/LayoutPage.tsx
@@ -1,33 +1,24 @@
-import { Application } from "@pixi/react";
-import React, { useState } from "react";
-import { CellMap, useCellMap } from "../game/CellMap";
+import React, { useCallback, useState } from "react";
+import { useCellMap } from "../game/CellMap";
 import {
   canPlaceMachine,
+  moveMachineAction,
+  placeMachineAction,
   removeMachineToStorageAction,
 } from "../game/game-actions/machine-actions";
+import { combineActions } from "../game/game-actions/misc-actions";
 import {
-  MACHINE_TYPES,
-  Machine,
-  MachineId,
-  MachineType,
-} from "../game/Machine";
+  addWorkItemAction,
+  applyWorkItemAction,
+} from "../game/game-actions/work-item-actions";
+import { GameState } from "../game/GameState";
+import { MACHINE_TYPES, MachineId, MachineType } from "../game/Machine";
 import { Direction } from "../game/Vectors";
-import { groupBy } from "../utils/arrayUtils";
-import { useTexture } from "../utils/useTexture";
+import { findPath } from "../utils/pathingUtils";
 import { NavBar } from "./NavBar";
-import { FloorTileSprite } from "./shop-view/FloorTileSprite";
-import { MachineSprite } from "./shop-view/MachineSprite";
-import { MaterialPilesSprite } from "./shop-view/MaterialPileSprite";
-import { PIXELS_PER_CELL, cellToPixel } from "./shop-view/shop-scale";
-import {
-  gameStateContext,
-  useApplyGameAction,
-  useGameState,
-  useLoadGame,
-  useMachines,
-  useNewGame,
-  useSaveGame,
-} from "./useGameState";
+import { LayoutEditorCanvas } from "./layout-page/LayoutEditorCanvas";
+import { StorageSection } from "./layout-page/StorageSection";
+import { useApplyGameAction, useGameState, useMachines } from "./useGameState";
 import { useKeyDown } from "./useKeyDown";
 
 interface PlacementMode {
@@ -53,6 +44,7 @@ export const LayoutPage: React.FC = () => {
   const gameState = useGameState();
   const machines = useMachines();
   const updateGameState = useApplyGameAction();
+  const cellMap = useCellMap();
 
   // Determine current edit mode
   const editMode: EditMode = placementMode
@@ -88,19 +80,105 @@ export const LayoutPage: React.FC = () => {
     }
   });
 
-  const saveGame = useSaveGame();
-  const loadGame = useLoadGame();
-  const newGame = useNewGame();
-  const cellMap = useCellMap();
-  const floorTexture = useTexture("/images/concrete-floor-2-big.png");
+  const handleFloorTileClick = useCallback(
+    (position: [number, number]) => {
+      // If in placement mode, try to place the machine
+      if (placementMode) {
+        const isValid = canPlaceMachine(
+          cellMap,
+          placementMode.machineType,
+          position,
+          placementMode.rotation,
+        );
 
-  const materialPileGroups = cellMap
-    .getCells()
-    .filter((cell) => cell.materialPiles.length > 0)
-    .map((cell) => cell.materialPiles);
+        if (isValid) {
+          updateGameState(
+            placeMachineAction(
+              placementMode.machineTypeId,
+              position,
+              placementMode.rotation,
+            ),
+          );
+          setPlacementMode(null);
+        }
+        return;
+      }
 
-  const width = cellToPixel(cellMap.getWidth());
-  const height = cellToPixel(cellMap.getHeight());
+      // If in move mode, try to move the selected machine
+      if (selectedMachineIndex !== null && selectedMachineIndex !== undefined) {
+        const selectedMachine = gameState.machines[selectedMachineIndex];
+        if (selectedMachine) {
+          const machineType = MACHINE_TYPES[selectedMachine.machineTypeId];
+          const isValid = canPlaceMachine(
+            cellMap,
+            machineType,
+            position,
+            moveRotation,
+            selectedMachineIndex, // Exclude the machine being moved
+          );
+
+          if (isValid) {
+            updateGameState(
+              moveMachineAction(selectedMachineIndex, position, moveRotation),
+            );
+            setSelectedMachineIndex(null);
+            setMoveRotation(0);
+          }
+        }
+        return;
+      }
+
+      // Normal player movement
+      const startPosition = getWorkQueueEndState(gameState).player.position;
+
+      const path = findPath(
+        startPosition,
+        position,
+        cellMap.getFreeCells().map((cell) => cell.position),
+      );
+
+      if (path != undefined) {
+        updateGameState(
+          combineActions(
+            ...path.map((pathItem) =>
+              addWorkItemAction({
+                type: "move",
+                direction: pathItem.direction,
+              }),
+            ),
+          ),
+        );
+      }
+    },
+    [
+      placementMode,
+      selectedMachineIndex,
+      moveRotation,
+      cellMap,
+      updateGameState,
+      gameState,
+    ],
+  );
+
+  const handleMachineClick = useCallback(
+    (index: number) => {
+      if (editMode === "none") {
+        // Toggle selection
+        if (selectedMachineIndex === index) {
+          setSelectedMachineIndex(null);
+        } else {
+          setSelectedMachineIndex(index);
+          setMoveRotation(gameState.machines[index].rotation);
+        }
+      }
+    },
+    [editMode, selectedMachineIndex, gameState.machines],
+  );
+
+  const handleBackgroundClick = useCallback(() => {
+    // Click on background deselects
+    setSelectedMachineIndex(null);
+  }, []);
 
   return (
     <main className="p-8 space-y-6">
@@ -108,96 +186,18 @@ export const LayoutPage: React.FC = () => {
 
       <div className="grid grid-cols-2">
         <section>
-          <Application
-            width={width}
-            height={height}
-            backgroundAlpha={0}
-            antialias={true}
-          >
-            <gameStateContext.Provider
-              value={{
-                gameState,
-                updateGameState,
-                saveGame,
-                loadGame,
-                newGame,
-              }}
-            >
-              <pixiContainer sortableChildren={true}>
-                <pixiTilingSprite
-                  eventMode="static"
-                  texture={floorTexture}
-                  tilePosition={{ x: 0, y: 0 }}
-                  tileScale={{ x: 0.25, y: 0.25 }}
-                  width={width}
-                  height={height}
-                  onClick={() => {
-                    // Click on background deselects
-                    setSelectedMachineIndex(null);
-                  }}
-                />
-                {cellMap.getCells().map((cell) => (
-                  <FloorTileSprite
-                    cell={cell}
-                    key={`cell-${cell.position.join(",")}`}
-                    placementMode={placementMode}
-                    selectedMachineIndex={selectedMachineIndex}
-                    moveRotation={moveRotation}
-                    onPlacementComplete={() => setPlacementMode(null)}
-                    onMoveComplete={() => {
-                      setSelectedMachineIndex(null);
-                      setMoveRotation(0);
-                    }}
-                    onHover={(position) => setHoverPosition(position)}
-                    onHoverOut={() => setHoverPosition(null)}
-                  />
-                ))}
-                {materialPileGroups.map((materialPiles, i) => (
-                  <pixiContainer
-                    key={`pile${materialPiles[0].position.join(",")}`}
-                    x={materialPiles[0].position[0] * PIXELS_PER_CELL}
-                    y={materialPiles[0].position[1] * PIXELS_PER_CELL}
-                  >
-                    <MaterialPilesSprite materialPiles={materialPiles} />
-                  </pixiContainer>
-                ))}
-                {machines.map((machinePlacement, index) => (
-                  <pixiContainer key={`machine-${index}`} zIndex={1000 + index}>
-                    <MachineSprite
-                      machine={machinePlacement}
-                      isSelected={selectedMachineIndex === index}
-                      onClick={() => {
-                        if (editMode === "none") {
-                          // Toggle selection
-                          if (selectedMachineIndex === index) {
-                            setSelectedMachineIndex(null);
-                          } else {
-                            setSelectedMachineIndex(index);
-                            setMoveRotation(machinePlacement.state.rotation);
-                          }
-                        }
-                      }}
-                    />
-                  </pixiContainer>
-                ))}
-                {placementMode && (
-                  <GhostMachinePreview
-                    placementMode={placementMode}
-                    cellMap={cellMap}
-                    hoverPosition={hoverPosition}
-                  />
-                )}
-                {selectedMachineIndex !== null && (
-                  <GhostMovePreview
-                    selectedMachineIndex={selectedMachineIndex}
-                    moveRotation={moveRotation}
-                    cellMap={cellMap}
-                    hoverPosition={hoverPosition}
-                  />
-                )}
-              </pixiContainer>
-            </gameStateContext.Provider>
-          </Application>
+          <LayoutEditorCanvas
+            cellMap={cellMap}
+            placementMode={placementMode}
+            selectedMachineIndex={selectedMachineIndex}
+            moveRotation={moveRotation}
+            hoverPosition={hoverPosition}
+            onFloorTileClick={handleFloorTileClick}
+            onHover={setHoverPosition}
+            onHoverOut={() => setHoverPosition(null)}
+            onMachineClick={handleMachineClick}
+            onBackgroundClick={handleBackgroundClick}
+          />
         </section>
         <StorageSection
           placementMode={placementMode}
@@ -211,231 +211,11 @@ export const LayoutPage: React.FC = () => {
   );
 };
 
-const StorageSection: React.FC<{
-  placementMode: PlacementMode | null;
-  setPlacementMode: (mode: PlacementMode | null) => void;
-  editMode: EditMode;
-  selectedMachineIndex: number | null;
-  machines: readonly any[];
-}> = ({
-  placementMode,
-  setPlacementMode,
-  editMode,
-  selectedMachineIndex,
-  machines,
-}) => {
-  const gameState = useGameState();
-  const groupedMachines = [
-    ...groupBy(gameState.storage.machines, (machineId) => machineId).values(),
-  ];
-
-  return (
-    <section className="space-y-4">
-      <h2 className="section-heading">Layout Editor</h2>
-
-      {/* Mode indicator */}
-      <div className="p-4 bg-brown-800 rounded-lg border-2 border-brown-700">
-        <div className="mb-2">
-          <span className="text-sm font-semibold text-brown-300">Mode: </span>
-          <span className="text-brown-100 font-bold">
-            {editMode === "placing" && "Placing Machine"}
-            {editMode === "moving" && "Moving Machine"}
-            {editMode === "none" && "Select or Place"}
-          </span>
-        </div>
-
-        {/* Keyboard shortcuts */}
-        <div className="text-xs text-brown-400 space-y-1">
-          {editMode === "placing" && (
-            <>
-              <div>• Click to place machine</div>
-              <div>• R to rotate</div>
-              <div>• ESC to cancel</div>
-            </>
-          )}
-          {editMode === "moving" && (
-            <>
-              <div>
-                • Moving:{" "}
-                {selectedMachineIndex !== null &&
-                  machines[selectedMachineIndex]?.type.name}
-              </div>
-              <div>• Click to move</div>
-              <div>• R to rotate</div>
-              <div>• Delete to remove</div>
-              <div>• ESC to cancel</div>
-            </>
-          )}
-          {editMode === "none" && (
-            <>
-              <div>• Click machine to select</div>
-              <div>• Click "Place" to add from storage</div>
-            </>
-          )}
-        </div>
-      </div>
-
-      {/* Storage inventory */}
-      <div>
-        <h3 className="text-lg font-semibold text-brown-100 mb-2">Storage</h3>
-        {groupedMachines.length === 0 ? (
-          <p className="text-brown-400 text-sm italic">
-            No machines in storage
-          </p>
-        ) : (
-          <ul className="space-y-2">
-            {groupedMachines.map((machineIds) => {
-              const machineType = MACHINE_TYPES[machineIds[0]];
-              const isSelected = placementMode?.machineTypeId === machineIds[0];
-              return (
-                <li
-                  key={machineIds[0]}
-                  className="flex items-center justify-between p-3 bg-brown-900 rounded border border-brown-700 hover:border-brown-600"
-                >
-                  <div>
-                    <div className="font-medium text-brown-100">
-                      {machineType.name}
-                    </div>
-                    <div className="text-xs text-brown-400">
-                      {machineType.description}
-                    </div>
-                  </div>
-                  <div className="flex items-center gap-2">
-                    {machineIds.length > 1 && (
-                      <span className="text-sm text-brown-300">
-                        ×{machineIds.length}
-                      </span>
-                    )}
-                    <button
-                      className={`button text-sm ${
-                        isSelected ? "bg-brown-600" : ""
-                      }`}
-                      onClick={() =>
-                        setPlacementMode({
-                          machineType,
-                          machineTypeId: machineIds[0],
-                          rotation: 0,
-                        })
-                      }
-                      disabled={editMode !== "none"}
-                    >
-                      {isSelected ? "Placing..." : "Place"}
-                    </button>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-        )}
-      </div>
-    </section>
-  );
-};
-
-const GhostMachinePreview: React.FC<{
-  placementMode: PlacementMode;
-  cellMap: CellMap;
-  hoverPosition: [number, number] | null;
-}> = ({ placementMode, cellMap, hoverPosition }) => {
-  const isValid =
-    hoverPosition && cellMap.has(hoverPosition)
-      ? canPlaceMachine(
-          cellMap,
-          placementMode.machineType,
-          hoverPosition,
-          placementMode.rotation,
-        )
-      : false;
-
-  // Create a temporary machine state for rendering
-  const ghostMachine = hoverPosition
-    ? {
-        machineTypeId: placementMode.machineTypeId,
-        position: hoverPosition,
-        rotation: placementMode.rotation,
-        selectedOperationId:
-          placementMode.machineType.operations.length > 0
-            ? placementMode.machineType.operations[0].id
-            : "none",
-        operationProgress: {
-          status: "notStarted" as const,
-          ticksRemaining: 0,
-        },
-        inputMaterials: [],
-        processingMaterials: [],
-        outputMaterials: [],
-      }
-    : null;
-
-  // Render ghost preview - non-interactive, just visual
-  if (!ghostMachine || !hoverPosition || !cellMap.has(hoverPosition)) {
-    return null;
+// Get the position at the end of the current work queue
+function getWorkQueueEndState(gameState: GameState): GameState {
+  let state = gameState;
+  for (const workItem of state.player.workQueue) {
+    state = applyWorkItemAction(workItem)(state);
   }
-
-  return (
-    <pixiContainer
-      alpha={0.6}
-      tint={isValid ? 0xffffff : 0xff6666}
-      eventMode="none"
-    >
-      <MachineSprite machine={new Machine(ghostMachine)} />
-    </pixiContainer>
-  );
-};
-
-const GhostMovePreview: React.FC<{
-  selectedMachineIndex: number;
-  moveRotation: Direction;
-  cellMap: CellMap;
-  hoverPosition: [number, number] | null;
-}> = ({ selectedMachineIndex, moveRotation, cellMap, hoverPosition }) => {
-  const gameState = useGameState();
-
-  const selectedMachine = gameState.machines[selectedMachineIndex];
-  if (!selectedMachine) return null;
-
-  const machineType = MACHINE_TYPES[selectedMachine.machineTypeId];
-
-  const isValid =
-    hoverPosition && cellMap.has(hoverPosition)
-      ? canPlaceMachine(
-          cellMap,
-          machineType,
-          hoverPosition,
-          moveRotation,
-          selectedMachineIndex,
-        )
-      : false;
-
-  // Create a temporary machine state for rendering
-  const ghostMachine = hoverPosition
-    ? {
-        machineTypeId: selectedMachine.machineTypeId,
-        position: hoverPosition,
-        rotation: moveRotation,
-        selectedOperationId: selectedMachine.selectedOperationId,
-        operationProgress: {
-          status: "notStarted" as const,
-          ticksRemaining: 0,
-        },
-        inputMaterials: [],
-        processingMaterials: [],
-        outputMaterials: [],
-      }
-    : null;
-
-  // Render ghost preview - non-interactive, just visual
-  if (!ghostMachine || !hoverPosition || !cellMap.has(hoverPosition)) {
-    return null;
-  }
-
-  return (
-    <pixiContainer
-      alpha={0.6}
-      tint={isValid ? 0xffffff : 0xff6666}
-      eventMode="none"
-    >
-      <MachineSprite machine={new Machine(ghostMachine)} />
-    </pixiContainer>
-  );
-};
+  return state;
+}
