+++ src/components/useGameState.tsx
@@ -1,4 +1,10 @@
-import React, { ReactNode, createContext, useContext, useState } from "react";
+import React, {
+  ReactNode,
+  createContext,
+  useContext,
+  useEffect,
+  useState,
+} from "react";
 import { GameState } from "../game/GameState";
 import { UpdateFunction } from "../utils/typeUtils";
 import { initialGameState } from "../game/initialGameState";
@@ -41,7 +47,7 @@ export const GameStateProvider: React.FC<{ children?: ReactNode }> = ({
   };
 
   // Expose game state functions to window for testing and debugging
-  React.useEffect(() => {
+  useEffect(() => {
     if (typeof window !== "undefined") {
       (window as any).__GAME_STATE__ = gameState;
       (window as any).__UPDATE_GAME_STATE__ = setGameState;
