+++ tests/fixtures/index.ts
@@ -1,10 +1,10 @@
 import { GameState } from "../../src/game/GameState";
-import layoutWithMiterSawInStorage from "./layout-with-miter-saw-in-storage.json";
-import layoutWithPlacedMachines from "./layout-with-placed-machines.json";
+import { layoutWithMiterSawInStorage } from "./layout-with-miter-saw-in-storage";
+import { layoutWithPlacedMachines } from "./layout-with-placed-machines";
 
 export const TEST_FIXTURES: Record<string, GameState> = {
-  "layout-with-miter-saw-in-storage": layoutWithMiterSawInStorage as unknown as GameState,
-  "layout-with-placed-machines": layoutWithPlacedMachines as unknown as GameState,
+  "layout-with-miter-saw-in-storage": layoutWithMiterSawInStorage,
+  "layout-with-placed-machines": layoutWithPlacedMachines,
 };
 
 // Expose fixtures to window for manual testing
