--- tests/fixtures/layout-with-placed-machines.json
@@ -1,81 +1,0 @@
-{
-  "tick": 0,
-  "money": 50,
-  "reputation": 2,
-  "materialPiles": [
-    {
-      "material": { "type": "board", "species": "oak", "width": 8, "length": 48, "thickness": 1 },
-      "position": [2, 1]
-    }
-  ],
-  "player": {
-    "name": "Player",
-    "position": [0, 0],
-    "direction": 0,
-    "inventory": [],
-    "workQueue": [],
-    "canWork": true
-  },
-  "machines": [
-    {
-      "machineTypeId": "workspace",
-      "position": [1, 2],
-      "rotation": 0,
-      "inputMaterials": [],
-      "processingMaterials": [],
-      "outputMaterials": [],
-      "selectedOperationId": "rip",
-      "selectedParameters": { "width": 8 },
-      "operationProgress": {
-        "status": "notStarted",
-        "ticksRemaining": 0
-      }
-    },
-    {
-      "machineTypeId": "miterSaw",
-      "position": [2, 4],
-      "rotation": 0,
-      "inputMaterials": [
-        { "type": "board", "species": "oak", "width": 8, "length": 48, "thickness": 1 }
-      ],
-      "processingMaterials": [],
-      "outputMaterials": [],
-      "selectedOperationId": "crosscut",
-      "selectedParameters": { "length": 24 },
-      "operationProgress": {
-        "status": "notStarted",
-        "ticksRemaining": 0
-      }
-    },
-    {
-      "machineTypeId": "makeshiftBench",
-      "position": [0, 4],
-      "rotation": 0,
-      "inputMaterials": [],
-      "processingMaterials": [],
-      "outputMaterials": [],
-      "selectedOperationId": "assemble",
-      "operationProgress": {
-        "status": "notStarted",
-        "ticksRemaining": 0
-      }
-    }
-  ],
-  "storage": {
-    "machines": []
-  },
-  "commissions": [],
-  "shopInfo": {
-    "name": "One Car Garage",
-    "electricity": 120,
-    "size": [4, 6],
-    "materialDropoffPosition": [3, 5]
-  },
-  "progression": {
-    "tutorialStage": 2,
-    "storeUnlocked": true,
-    "shopLayoutUnlocked": true,
-    "freeSelling": false,
-    "commissionsCompleted": 1
-  }
-}
