+++ tsconfig.json
@@ -6,8 +6,7 @@
     "moduleResolution": "node",
     "jsx": "react",
     "esModuleInterop": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true
+    "skipLibCheck": true
   }
   // "include": ["src/types", "src/server/index.ts", "src/client/pages"]
 }
